pyproject.toml
[project]
name = "tennis-data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "dbt-bigquery>=1.11.1",
    "dbt-core>=1.11.7",
    "google-cloud-bigquery",
    "google-cloud-storage",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "apache-airflow>=3.1.8",
]

[tool.uv]
exclude-newer = "7d"
